Portfolio Website
This is my personal portfolio website showcasing my skills, projects, and experience as a web developer. It features a clean and responsive design, with sections highlighting my background, technical expertise, and accomplishments. The site is built using HTML, CSS, and JavaScript, and is fully responsive, ensuring a seamless experience across all devices.